requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
dotenv==0.9.9
Faker==37.12.0
num2words==0.5.14
pillow==12.1.1
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.3
tzdata==2025.2
python-dotenv
psycopg2-binary>=2.9
gunicorn
